Update package.xml, launch script for package split
