--- launch/hobot_body_tracking.launch.py
@@ -1,119 +1,0 @@
-# Copyright (c) 2022，Horizon Robotics.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-
-from launch import LaunchDescription
-from launch_ros.actions import Node
-
-from launch.actions import IncludeLaunchDescription
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-from ament_index_python import get_package_share_directory
-
-def generate_launch_description():
-    web_service_launch_include = IncludeLaunchDescription(
-        PythonLaunchDescriptionSource(
-            os.path.join(
-                get_package_share_directory('websocket'),
-                'launch/hobot_websocket_service.launch.py'))
-    )
-
-    return LaunchDescription([
-        web_service_launch_include,
-        # 启动图片发布pkg
-        Node(
-            package='mipi_cam',
-            executable='mipi_cam',
-            output='screen',
-            parameters=[
-                {"out_format": "nv12"},
-                {"image_width": 960},
-                {"image_height": 544},
-                {"io_method": "shared_mem"},
-                {"video_device": "F37"}
-            ],
-            arguments=['--ros-args', '--log-level', 'error']
-        ),
-        # 启动jpeg图片编码&发布pkg
-        Node(
-            package='hobot_codec',
-            executable='hobot_codec_republish',
-            output='screen',
-            parameters=[
-                {"channel": 1},
-                {"in_mode": "shared_mem"},
-                {"in_format": "nv12"},
-                {"out_mode": "ros"},
-                {"out_format": "jpeg"},
-                {"sub_topic": "/hbmem_img"},
-                {"pub_topic": "/image_jpeg"}
-            ],
-            arguments=['--ros-args', '--log-level', 'error']
-        ),
-        # 启动单目rgb人体、人头、人脸、人手框和人体关键点检测pkg
-        Node(
-            package='mono2d_body_detection',
-            executable='mono2d_body_detection',
-            output='screen',
-            parameters=[
-            ],
-            arguments=['--ros-args', '--log-level', 'error']
-        ),
-        # 启动人手关键点检测pkg
-        Node(
-            package='hand_lmk_detection',
-            executable='hand_lmk_detection',
-            output='screen',
-            parameters=[
-            ],
-            arguments=['--ros-args', '--log-level', 'error']
-        ),
-        # 启动web展示pkg
-        Node(
-            package='websocket',
-            executable='websocket',
-            output='screen',
-            parameters=[
-                {"image_topic": "/image_jpeg"},
-                {"image_type": "mjpeg"},
-                {"smart_topic": "/hobot_hand_gesture_detection"}
-            ],
-            arguments=['--ros-args', '--log-level', 'error']
-        ),
-        # 启动手势识别pkg
-        Node(
-            package='hand_gesture_detection',
-            executable='hand_gesture_detection',
-            output='screen',
-            parameters=[
-            ],
-            arguments=['--ros-args', '--log-level', 'error']
-        ),
-        # 启动人体跟随pkg
-        Node(
-            package='body_tracking',
-            executable='body_tracking',
-            output='screen',
-            parameters=[
-                {"activate_wakeup_gesture": 0},
-                {"img_width": 960},
-                {"img_height": 544},
-                {"track_serial_lost_num_thr": 30},
-                {"move_step": 0.5},
-                {"rotate_step": 0.5},
-                {"activate_robot_move_thr": 5}
-            ],
-            arguments=['--ros-args', '--log-level', 'info']
-        )
-    ])
