+++ package.xml
@@ -2,8 +2,8 @@
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
   <name>body_tracking</name>
-  <version>0.0.0</version>
-  <description>body tracking</description>
+  <version>1.1.6</version>
+  <description>TogetheROS body tracking</description>
   <maintainer email="kao.zhu@horizon.ai">kao.zhu</maintainer>
   <license>Apache License 2.0</license>
 
@@ -17,6 +17,9 @@
 
   <depend>rclcpp_action</depend>
   <depend>geometry_msgs</depend>
+
+  <exec_depend>hand_gesture_detection</exec_depend>
+  <exec_depend>mono2d_body_detection</exec_depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
